requirements.txt
streamlit
python-dotenv
praw